% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbl_tbl.R
\name{as_lbl_tbl}
\alias{as_lbl_tbl}
\title{as labeled table}
\usage{
as_lbl_tbl(x, ...)
}
\arguments{
\item{x}{a tibble or data frame}

\item{...}{\code{name = value} pairs (or a list) of data frame labels}
}
\value{
a labeled tibble (\code{lbl_tbl}) object
}
\description{
Convert a data frame or tibble into a labeled table (\code{lbl_tbl}) by providing metadata labels.
Metadata labels can include \code{name}, \code{title}, \code{version}, \code{description}, \code{created}.
By default, \code{x} is deparsed to create the \code{name} label and the current system time is used as the \code{created} label.
}
\examples{
as_lbl_tbl(mtcars, name = "mtcars", title = "Motor Trend Road Car Tests")
d <-
  tibble::tibble(
    id = labels(letters[1:5]),
    letters = as_lbl_vec(letters[1:5], "The Alphabet"),
    numbers = as_lbl_vec(1:5, "the numbers"),
    logicals = as_lbl_vec(c(TRUE, TRUE, FALSE, TRUE, FALSE), "is foofy")
  )
d <- as_lbl_tbl(d)
d$letters
}
