% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbl_vec.R
\name{as_lbl_vec}
\alias{as_lbl_vec}
\title{as labeled vector}
\usage{
as_lbl_vec(x = vector(), label, ...)
}
\arguments{
\item{x}{a vector (see \code{vctrs::obj_check_vector()})
type \code{lbl_vec}, a labeled vector}
}
\description{
Convert a vector into a labeled vector (\code{lbl_vec}) by providing a label
}
\examples{
as_lbl_vec(letters, "the alphabet")
as_lbl_vec(1:10, "the numbers")
as_lbl_vec(c(T, T, F, T), "is foofy")
x <- as_lbl_vec(letters, "The Alphabet")
}
