% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{as_markdown}
\alias{as_markdown}
\title{Generate markdown representation of a list}
\usage{
as_markdown(x, indent = 0)
}
\arguments{
\item{x}{(list) Object.}

\item{indent}{(integer) Base indentation level of markdown list.}
}
\description{
Generate markdown representation of a list
}
\seealso{
Other package writers: \code{\link{write_data}},
  \code{\link{write_meta}}, \code{\link{write_package}},
  \code{\link{write_resource}}
}
