% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{compress_file}
\alias{compress_file}
\title{Compress file}
\usage{
compress_file(file, ext = c("gz", "bz2"), remove = FALSE)
}
\arguments{
\item{file}{(character) Path to file.}

\item{ext}{(character) Compression file extension.}

\item{remove}{(logical) Whether to remove the original file.}
}
\description{
Compress file
}
