% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_objects.R
\name{constraints}
\alias{constraints}
\title{Create field constraints}
\usage{
constraints(required = NULL, unique = NULL, minLength = NULL,
  maxLength = NULL, minimum = NULL, maximum = NULL, pattern = NULL,
  enum = NULL)
}
\arguments{
\item{required}{(logical) Whether field cannot contain \code{NA}.}

\item{unique}{(logical) Whether field values must be unique.}

\item{minLength}{(integer) Minimum length of field values.}

\item{maxLength}{(integer) Maximum length of field values.}

\item{minimum}{Minimum value (of same type as field).}

\item{maximum}{Maximum value (of same type as field).}

\item{pattern}{(character) Regular expression field values must match (see \url{https://www.w3.org/TR/xmlschema-2/#regexs}).}

\item{enum}{(vector) Allowed values.}
}
\description{
A list of constraints for validating field values.
}
\examples{
constraints(
  required = TRUE,
  unique = TRUE,
  minimum = "1970-01-01",
  pattern = "[0-9]{4}-[0-9]{2}-[0-9]{2}"
)
}
\references{
\url{https://specs.frictionlessdata.io/table-schema/#constraints}
}
\seealso{
Other meta objects: \code{\link{contributor}},
  \code{\link{field}}, \code{\link{foreignKey}},
  \code{\link{package}}, \code{\link{resource}},
  \code{\link{schema}}, \code{\link{source}}
}
