% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{decompress_file}
\alias{decompress_file}
\title{Decompress file}
\usage{
decompress_file(file, remove = FALSE)
}
\arguments{
\item{file}{(character) Path to compressed file.}

\item{remove}{(logical) Whether to remove the original file.}
}
\description{
Decompress file
}
