% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deparsers.R
\name{deparse_field}
\alias{deparse_field}
\alias{deparse_string}
\alias{deparse_number}
\alias{deparse_integer}
\alias{deparse_boolean}
\alias{deparse_date}
\alias{deparse_datetime}
\title{Deparse field}
\usage{
deparse_field(x, meta = get_field(x))

deparse_string(x)

deparse_number(x)

deparse_integer(x)

deparse_boolean(x)

deparse_date(x)

deparse_datetime(x)
}
\arguments{
\item{x}{(atomic) Object to deparse.}

\item{meta}{(list) Field metadata.}
}
\description{
Deparse field
}
\details{
NOTES:
\itemize{
  \item Ignores missingValues (leaves to write_* functions)
  \item In some cases, leaves final coercion to write_* functions
}
}
\examples{
deparse_field(Sys.Date())
deparse_field(Sys.time())
deparse_field("string")
deparse_field(1)
}
