% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deparsers.R
\name{deparse_fields}
\alias{deparse_fields}
\title{Deparse list of fields}
\usage{
deparse_fields(l, meta = get_fields(l))
}
\arguments{
\item{l}{(list of atomic vectors)}

\item{meta}{(list of lists) Field metadata}
}
\description{
Deparse list of fields
}
