% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_dpkg.R
\name{download_dpkg}
\alias{download_dpkg}
\title{download a file to the dpkg R user directory}
\usage{
download_dpkg(uri)
}
\description{
Files are downloaded to the R user data
directory and are cached across all of an R user’s sessions and projects.
After reading an online data package for the first time, it will be cached
has been read once it will not need to be downloaded again
when it is read again.
Specify an alternative download location by setting the \code{R_USER_DATA_DIR}
environment variable; see \code{?tools::R_user_dir}.
}
