% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpkg_release_gh.R
\name{dpkg_badge}
\alias{dpkg_badge}
\title{create readme badge for a dpkg github release}
\usage{
dpkg_badge(release_url)
}
\arguments{
\item{release_url}{the url of a github release (as returned by \code{dpkg_gh_release()})}
}
\value{
character string of markdown
}
\description{
The badge relies on shields.io for the images, which will always
point to the most recently released version.
Note that this relies on the structure of the release created with
dpkg_gh_release().
}
\examples{
release_url <- "https://github.com/cole-brokamp/dpkg/releases/tag/mtcars-v0.0.0.9000"
dpkg_badge(release_url)
dpkg_badge("https://github.com/geomarker-io/parcel/releases/tag/property_code_enforcements-v1.0.0")
dpkg_badge("https://github.com/geomarker-io/parcel/releases/tag/cagis_parcels-v1.1.0")
dpkg_badge("https://github.com/geomarker-io/parcel/releases/tag/auditor_online_parcels-v0.2.0")
}
