% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_lbl_tbl.R
\name{example_lbl_tbl}
\alias{example_lbl_tbl}
\title{example labeled table (\code{lbl_tbl})}
\usage{
example_lbl_tbl()
}
\description{
example labeled table (\code{lbl_tbl})
}
\examples{
example_lbl_tbl()
}
