% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_objects.R
\name{field}
\alias{field}
\title{Create field metadata}
\usage{
field(name = NULL, type = "string", format = NULL, title = NULL,
  description = NULL, rdfType = NULL, constraints = list(), unit = NULL,
  ...)
}
\arguments{
\item{name}{(character) Field name.}

\item{type}{(character) Field type (see \href{https://specs.frictionlessdata.io/table-schema/#types-and-formats}{table-schema/types-and-formats}).}

\item{format}{(character) Field format (see \href{https://specs.frictionlessdata.io/table-schema/#types-and-formats}{table-schema/types-and-formats}).}

\item{title}{(character) Human-readable title.}

\item{description}{(character) Human-readable description.}

\item{rdfType}{(character) URI of an RDF class (see \href{https://specs.frictionlessdata.io/table-schema/#rich-types}{table-schema/rich-types}).}

\item{constraints}{(named list) See \code{\link{constraints}}.}

\item{unit}{(character) Unit of measurement.}

\item{...}{(name = value) Additional properties.}
}
\description{
The metadata for a field. See \code{\link{set_field}} and \code{\link{get_field}} to set and get field metadata to and from an object.
}
\examples{
field(
 name = "date_created",
 type = "date",
 format = "\%Y-\%m-\%d"
)
}
\references{
\url{https://specs.frictionlessdata.io/table-schema/#field-descriptors}
}
\seealso{
Other meta objects: \code{\link{constraints}},
  \code{\link{contributor}}, \code{\link{foreignKey}},
  \code{\link{package}}, \code{\link{resource}},
  \code{\link{schema}}, \code{\link{source}}

Other field functions: \code{\link{get_field}},
  \code{\link{set_fields}}, \code{\link{set_field}}
}
