% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_objects.R
\name{foreignKey}
\alias{foreignKey}
\title{Create foreign key}
\usage{
foreignKey(fields, resource, resource_fields)
}
\arguments{
\item{fields}{(character) Local field name(s).}

\item{resource}{(character) Foreign resource name (within the same data package) or \code{""} for the local resource.}

\item{resource_fields}{(character) Foreign field name(s). Must match length of \code{fields}.}
}
\description{
A reference between fields in different data package resources (tables).
}
\examples{
foreignKey(
  fields = "state-code",
  resource = "state-codes",
  resource_fields = "code"
)
}
\references{
\url{https://specs.frictionlessdata.io/table-schema/#foreign-keys}
}
\seealso{
Other meta objects: \code{\link{constraints}},
  \code{\link{contributor}}, \code{\link{field}},
  \code{\link{package}}, \code{\link{resource}},
  \code{\link{schema}}, \code{\link{source}}
}
