% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\name{get_field}
\alias{get_field}
\alias{get_fields}
\title{Get field metadata}
\usage{
get_field(x, name = NULL)

get_fields(l)
}
\arguments{
\item{x}{Object.}

\item{name}{(character) Field name to return if not set explicitly.}

\item{l}{(list) List of objects.}
}
\description{
Missing properties are returned filled with their default values (see Details).
}
\details{
Unless set explicity, the following defaults are returned:
\itemize{
  \item name: The name of the object in a list.
  \item type: The type corresponding to the object class (or "string" if not supported).
  \item format: The default format for that type.
  \item unit: Units set by \code{\link[units]{units}} deparsed to product power form (if available).
}
}
\examples{
get_field(1)
get_field(1, "id")
x <- set_field(1, type = "logical")
get_field(x)
get_field(Sys.time())
get_fields(data.frame(id = 1L, value = 1.1))
}
\seealso{
Other field functions: \code{\link{field}},
  \code{\link{set_fields}}, \code{\link{set_field}}
}
