% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\name{get_field_format_default}
\alias{get_field_format_default}
\title{Get default format by field type}
\usage{
get_field_format_default(type)
}
\arguments{
\item{type}{(character) Field type.}
}
\description{
Get default format by field type
}
