% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbl_vec.R
\name{get_label}
\alias{get_label}
\title{get label}
\usage{
get_label(x, missing_label = c("na", "deparse", "as_label"))
}
\arguments{
\item{x}{a \code{lbl_vec} object (see \code{as_lbl_vec()})}

\item{missing_label}{what to return in the case of no label attribute: \code{"na"} returns missing values,
\code{"deparse"} substitutes and deparses the code used to define \code{x} into a character string, and
\code{"as_label"} tries to convert the code supplied as \code{x} to a label with \code{rlang::as_label()}}
}
\value{
character string of label or \code{NA} if label does not exist
}
\description{
get the label of a labeled vector
}
\examples{
as_lbl_vec(letters, "The Letters") |>
  get_label()
get_label(letters)
get_label(letters, missing_label = "deparse")
get_label(letters, missing_label = "as_label")
toupper(letters) |>
  get_label(missing_label = "deparse")
toupper(letters) |>
  get_label(missing_label = "as_label")
}
