% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbl_tbl.R
\name{get_md}
\alias{get_md}
\title{get metadata}
\usage{
get_md(x, which = c("name", "title", "version", "description", "created"))
}
\arguments{
\item{x}{a \code{lbl_tbl} object (see \code{as_lbl_tbl()})}

\item{which}{character vector of which labels to return}
}
\value{
a list of table-specific labels
}
\description{
Get table-specific labels (i.e. metadata) associated with a labeled table (\code{lbl_tbl}) object.
}
\examples{
d_md <- get_md(example_lbl_tbl())
d_md[c("name", "title")]
str(d_md)
}
