% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbl_tbl.R
\name{get_meta}
\alias{get_meta}
\title{get metadata}
\usage{
get_meta(x, which = c("name", "title", "version", "description", "created"))
}
\arguments{
\item{x}{a \code{lbl_tbl} object (see \code{as_lbl_tbl()})}

\item{which}{character vector of which metadata values to return}
}
\value{
a list of table-specific metadata
}
\description{
Get table-specific metadata associated with a labeled table (\code{lbl_tbl}) object.
}
\examples{
d_md <- get_meta(example_lbl_tbl())
d_md[c("name", "title")]
str(d_md)
}
