% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\name{get_package}
\alias{get_package}
\title{Get package metadata}
\usage{
get_package(x, name = NULL, inline_data = TRUE)
}
\arguments{
\item{x}{Object.}

\item{name}{(character) Package name to return if not set explicitly.}

\item{inline_data}{(logical) Whether to include the contents of \code{x} as inline data (\code{$data}).}
}
\description{
Missing properties are returned filled with their default values (see Details). An error is returned if the resources property is empty.
}
\details{
Unless set explicity, the following defaults are returned:
\itemize{
  \item resources: Resource metadata from the elements of the object, via \code{\link{get_resource}}.
}
}
\examples{
x <- list(data = data.frame(id = 1L, value = 1.1))
get_package(x, inline_data = FALSE)
}
\seealso{
Other resource functions: \code{\link{get_resource}},
  \code{\link{resource}}, \code{\link{set_resources}},
  \code{\link{set_resource}}
}
