% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\name{get_resource}
\alias{get_resource}
\alias{get_resources}
\title{Get resource metadata}
\usage{
get_resource(x, name = NULL, inline_data = TRUE)

get_resources(l, inline_data = FALSE)
}
\arguments{
\item{x}{Object.}

\item{name}{(character) Resource name to return if not set explicitly.}

\item{inline_data}{(logical) Whether to include the contents of \code{x} as inline data (\code{$data}).}

\item{l}{(list) List of objects.}
}
\description{
Missing properties are returned filled with their default values (see Details). If \code{inline_data = TRUE} and property \code{path} is \code{NULL}, object data is returned in the \code{data} property either unchanged (if \code{format} is missing or "json") or formatted as a string (provided the format is supported).
}
\details{
Unless set explicity, the following defaults are returned:
\itemize{
  \item name: The name of the object in a list.
  \item schema$fields: Field metadata from the elements of the object, via \code{\link{get_field}}.
}
}
\examples{
x <- data.frame(id = 1L, value = 1.1)
get_resource(x)
get_resource(x, "data")
l <- set_resources(
  list(a = data.frame(), b = data.frame(x = 1)),
  a = resource(path = "data/a.csv"),
  b = resource(path = "data/b.csv")
)
get_resources(l)
}
\seealso{
Other resource functions: \code{\link{get_package}},
  \code{\link{resource}}, \code{\link{set_resources}},
  \code{\link{set_resource}}
}
