% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{github_raw_url}
\alias{github_raw_url}
\title{Convert GitHub repository address to raw url}
\usage{
github_raw_url(repo)
}
\arguments{
\item{repo}{Repository address in the format \code{username/repo[/subdir][@ref]}, where \code{ref} can be a commit, tag, or branch name. Defaults to "master".}
}
\description{
Convert GitHub repository address to raw url
}
