% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{is_compressed}
\alias{is_compressed}
\title{Test if path is for a compressed file}
\usage{
is_compressed(file)
}
\arguments{
\item{file}{(character) Path to file.}
}
\description{
Test if path is for a compressed file
}
