% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_empty}
\alias{is_empty}
\title{Check whether an object is empty or contains NA or ""}
\usage{
is_empty(x)
}
\arguments{
\item{x}{Object.}
}
\description{
Check whether an object is empty or contains NA or ""
}
