% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_list_not_df}
\alias{is_list_not_df}
\title{Check whether an object is a list (but not a data.frame)}
\usage{
is_list_not_df(x)
}
\arguments{
\item{x}{Object.}
}
\description{
Check whether an object is a list (but not a data.frame)
}
