% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_objects.R
\name{license}
\alias{license}
\title{Create license}
\usage{
license(name = NULL, path = NULL, title = NULL)
}
\arguments{
\item{name}{(character) Open Definition license ID (see \url{http://licenses.opendefinition.org/}). Required if \code{path} is \code{NULL}.}

\item{path}{(character) URL or path (see \href{https://specs.frictionlessdata.io/data-resource/#url-or-path}{data-resource/url-or-path}). Required if \code{name} is \code{NULL}.}

\item{title}{(character) Human-readable title.}
}
\description{
A license under which the data package is provided.
}
\examples{
license(
  name = "ODC-PDDL-1.0",
  path = "http://opendatacommons.org/licenses/pddl/",
  title = "Open Data Commons Public Domain Dedication and License v1.0"
)
}
\references{
\url{https://specs.frictionlessdata.io/data-package/#licenses}
}
