% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{modify_attr}
\alias{modify_attr}
\title{Modify object list attribute
TODO: Set attribute by reference.}
\usage{
modify_attr(x, which, value)
}
\arguments{
\item{x}{Object.}

\item{which}{(character) Name of attribute to modify.}

\item{value}{(list) New attribute value.}
}
\description{
Modify object list attribute
TODO: Set attribute by reference.
}
