% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_objects.R
\name{package}
\alias{package}
\title{Create package metadata}
\usage{
package(name = NULL, title = NULL, description = NULL, homepage = NULL,
  id = NULL, profile = NULL, version = NULL, created = NULL,
  sources = list(), contributors = list(), licenses = list(),
  keywords = NULL, image = NULL, resources = list(), ...)
}
\arguments{
\item{name}{(character) Package name.}

\item{title}{(character) Human-readable title.}

\item{description}{(character) Human-readable description.}

\item{homepage}{(character) URL of the package homepage.}

\item{id}{(character) Globally unique identifier.}

\item{profile}{(character) Custom package profile (see \url{https://specs.frictionlessdata.io/profiles/}).}

\item{version}{(character) Version specification (see \url{http://semver.org/}).}

\item{created}{(character) ISO 8601 date and time of package creation (see \url{https://tools.ietf.org/html/rfc3339#section-5.6}).}

\item{sources}{(list) See \code{\link{source}}.}

\item{contributors}{(list) See \code{\link{contributor}}.}

\item{licenses}{(list) See \code{\link{license}}.}

\item{keywords}{(character) Keywords describing the package.}

\item{image}{(character) URL or path relative to the package directory (see \href{https://specs.frictionlessdata.io/data-resource/#url-or-path}{data-resource/url-or-path}).}

\item{resources}{(list) See \code{\link{resource}}.}

\item{...}{(name = value) Additional properties.}
}
\description{
The metadata for a data package. See \code{\link{set_package}} and \code{\link{get_package}} to set and get package metadata to and from an object.
}
\details{
TODO:
\itemize{
  \item Require name?
  \item Name must be lower-case and contain only alphanumeric characters and \code{.}, \code{_}, and \code{-}.
}
}
\examples{
package(
 name = "example",
 title = "Example data package"
)
}
\references{
\url{https://specs.frictionlessdata.io/data-package/}
}
\seealso{
Other meta objects: \code{\link{constraints}},
  \code{\link{contributor}}, \code{\link{field}},
  \code{\link{foreignKey}}, \code{\link{resource}},
  \code{\link{schema}}, \code{\link{source}}

Other package functions: \code{\link{set_package}}
}
