% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_field}
\alias{parse_field}
\alias{parse_string}
\alias{parse_number}
\alias{parse_integer}
\alias{parse_boolean}
\alias{parse_date}
\alias{parse_datetime}
\title{Parse field}
\usage{
parse_field(x, meta = NULL)

parse_string(x)

parse_number(x, decimalChar = ".", groupChar = NULL, bareNumber = TRUE,
  unit = NULL)

parse_integer(x, bareNumber = TRUE, unit = NULL)

parse_boolean(x, trueValues = c("true", "True", "TRUE", "1"),
  falseValues = c("false", "False", "FALSE", "0"))

parse_date(x, format = "\%Y-\%m-\%d")

parse_datetime(x, format = "\%Y-\%m-\%dT\%H:\%M:\%SZ")
}
\arguments{
\item{x}{(atomic) Object to parse.}

\item{meta}{(named list) Field metadata.}

\item{decimalChar}{(character) Symbol used to indicate the decimal place.}

\item{groupChar}{(character) Symbol used to chunk larger numbers.}

\item{bareNumber}{(boolean) Whether \code{x}, if character, follows the formatting constraints of a number. If \code{FALSE}, trailing non-numeric characters are first removed.}

\item{unit}{(character) Unit of measure in product power form (see \code{\link[units]{parse_unit}}).}

\item{trueValues}{(character) Values indicating \code{TRUE}.}

\item{falseValues}{(character) Values indicating \code{FALSE}.}

\item{format}{(character) Format specification (see \code{\link[base]{strptime}}).}
}
\description{
Parse field
}
\details{
TODO:
\itemize{
  \item string: Support format property
  \item number: Support currency property
  \item Validate field against constraints
}
}
\references{
\url{http://specs.frictionlessdata.io/table-schema/#field-descriptors}
}
\seealso{
Other field parsers: \code{\link{parse_fields}}
}
