% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_fields}
\alias{parse_fields}
\title{Parse list of fields}
\usage{
parse_fields(l, meta = list())
}
\arguments{
\item{l}{(list) Atomic vectors to parse.}

\item{meta}{(list of named lists) Field metadata.}
}
\description{
Parse list of fields
}
\references{
\url{http://specs.frictionlessdata.io/tabular-data-resource/}
}
\seealso{
Other field parsers: \code{\link{parse_field}}
}
