% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_resource}
\alias{parse_resource}
\title{Parse resource}
\usage{
parse_resource(meta, path = getwd())
}
\arguments{
\item{meta}{(list) Resource metadata.}

\item{path}{(character) Path to data package.}
}
\description{
Parse resource
}
\details{
NOTES:
\itemize{
  \item Assumes tabular data if the schema contains one or more fields.
  \item Only supports json and csv.
}
}
\references{
\url{https://specs.frictionlessdata.io/data-resource/},
\url{https://specs.frictionlessdata.io/tabular-data-resource/}
}
\seealso{
Other resource parsers: \code{\link{parse_resource_data}},
  \code{\link{parse_resource_format}}
}
