% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_resource_data}
\alias{parse_resource_data}
\title{Parse resource data}
\usage{
parse_resource_data(x, meta)
}
\arguments{
\item{x}{(character, list) Object to parse.}

\item{meta}{(named list) Resource metadata.}
}
\description{
Parse resource data
}
\references{
\url{https://specs.frictionlessdata.io/data-resource/}
}
\seealso{
Other resource parsers: \code{\link{parse_resource_format}},
  \code{\link{parse_resource}}
}
