% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_resource_format}
\alias{parse_resource_format}
\title{Parse resource format}
\usage{
parse_resource_format(meta)
}
\arguments{
\item{meta}{(character) Resource metadata.}
}
\description{
Parse resource format
}
\references{
\url{https://specs.frictionlessdata.io/data-resource/}
}
\seealso{
Other resource parsers: \code{\link{parse_resource_data}},
  \code{\link{parse_resource}}
}
