% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{read_csv}
\alias{read_csv}
\title{Read CSV}
\usage{
read_csv(txt, na = NULL, colClasses = NULL)
}
\arguments{
\item{txt}{(character) CSV string or path.}

\item{na}{(character) Strings interpreted as \code{NA}.}

\item{colClasses}{(character) Classes to assume for the columns.}
}
\description{
TODO: Distinguish between empty and empty quoted string.
}
\seealso{
Other formats: \code{\link{read_json}},
  \code{\link{write_csv}}, \code{\link{write_json}}
}
