% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_data}
\alias{read_data}
\alias{read_data_github}
\title{Read package data}
\usage{
read_data(path = getwd(), resources = NULL, meta = read_meta(path),
  drop = FALSE)

read_data_github(repo, ...)
}
\arguments{
\item{path}{(character) Package directory or URL.}

\item{resources}{(character, numeric) Names or indices of resources to read (or all if \code{NULL}).}

\item{meta}{(list) Package metadata.}

\item{drop}{(logical) Whether to unlist a single resource.}

\item{repo}{(character) GitHub repository address (see \code{\link{github_raw_url}}).}

\item{...}{Additional arguments passed to \code{\link{read_data}}.}
}
\description{
Read package data
}
\references{
\url{https://specs.frictionlessdata.io/data-resource/}
}
\seealso{
Other package readers: \code{\link{read_meta}},
  \code{\link{read_package}}
}
