% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_dpkg.R
\name{read_dpkg_metadata}
\alias{read_dpkg_metadata}
\alias{read_dpkg}
\title{read (meta)data from dpkg on disk}
\usage{
read_dpkg_metadata(x)

read_dpkg(x)
}
\arguments{
\item{x}{path to data package (\code{.parquet} file) on disk}
}
\value{
for \code{read_dpkg()}, a dpkg object; for \code{read_dpkg_metadata()}, a list of metadata
}
\description{
read (meta)data from dpkg on disk
}
