% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_dpkg.R
\name{read_dpkg_metadata}
\alias{read_dpkg_metadata}
\title{read metadata from dpkg on disk}
\usage{
read_dpkg_metadata(x)
}
\arguments{
\item{x}{path to data package (\code{.parquet} file) on disk}
}
\value{
a list of metadata collected from parquet_info and the key value metadata
}
\description{
read metadata from dpkg on disk
}
