% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{read_json}
\alias{read_json}
\title{Read JSON}
\usage{
read_json(txt, tabular = FALSE)
}
\arguments{
\item{txt}{(character) JSON string or path.}

\item{tabular}{(logical) Whether to read \code{txt} as tabular data.}
}
\description{
Read JSON
}
\references{
\url{https://specs.frictionlessdata.io/tabular-data-resource/#json-tabular-data}
}
\seealso{
Other formats: \code{\link{read_csv}},
  \code{\link{write_csv}}, \code{\link{write_json}}
}
