% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_meta}
\alias{read_meta}
\alias{read_meta_github}
\title{Read package metadata}
\usage{
read_meta(path = getwd())

read_meta_github(repo)
}
\arguments{
\item{path}{(character) Package directory or URL.}

\item{repo}{(character) GitHub repository address (see \code{\link{github_raw_url}}).}
}
\description{
Read package metadata
}
\seealso{
Other package readers: \code{\link{read_data}},
  \code{\link{read_package}}
}
