% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_objects.R
\name{resource}
\alias{resource}
\title{Create resource metadata}
\usage{
resource(name = NULL, path = NULL, profile = NULL, title = NULL,
  description = NULL, format = NULL, mediatype = NULL,
  encoding = "UTF-8", schema = list(), bytes = NULL, hash = NULL,
  sources = list(), licenses = list(), ...)
}
\arguments{
\item{name}{(character) Resource name.}

\item{path}{(character) URLs or paths relative to the data package directory (see \href{https://specs.frictionlessdata.io/data-resource/#path-data-in-files}{data-resource/path-data-in-files}).}

\item{profile}{(character) Custom resource profile (see \url{https://specs.frictionlessdata.io/profiles/}).}

\item{title}{(character) Human-readable title.}

\item{description}{(character) Human-readable description.}

\item{format}{(character) Output format (e.g. "csv", "json").}

\item{mediatype}{(character) Mediatype/mimetype of the output (e.g. "text/csv", "application/json").}

\item{encoding}{(character) Character encoding of the output (see \url{http://www.iana.org/assignments/character-sets/character-sets.xhtml}).}

\item{schema}{(named list) See \code{\link{schema}}.}

\item{bytes}{(integer) Size of the output file in bytes.}

\item{hash}{(character) MD5 hash of the output file.}

\item{sources}{(list) See \code{\link{source}}.}

\item{licenses}{(list) See \code{\link{license}}.}

\item{...}{(name = value) Additional properties.}
}
\description{
The metadata for a data resource. See \code{\link{set_resource}} and \code{\link{get_resource}} to set and get resource metadata to and from an object.
}
\examples{
resource(
 name = "data",
 path = "data/data.csv",
 format = "csv",
 schema = schema(missingValues = "")
)
}
\references{
\url{https://specs.frictionlessdata.io/data-resource/}
}
\seealso{
Other meta objects: \code{\link{constraints}},
  \code{\link{contributor}}, \code{\link{field}},
  \code{\link{foreignKey}}, \code{\link{package}},
  \code{\link{schema}}, \code{\link{source}}

Other resource functions: \code{\link{get_package}},
  \code{\link{get_resource}}, \code{\link{set_resources}},
  \code{\link{set_resource}}
}
