% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_objects.R
\name{schema}
\alias{schema}
\title{Create table schema}
\usage{
schema(fields = list(), missingValues = NULL, primaryKey = NULL,
  foreignKeys = list(), ...)
}
\arguments{
\item{fields}{(list of named lists) See \code{\link{field}}. An empty \code{fields} is permitted to allow use with a \code{\link{resource}} data object.}

\item{missingValues}{(character) Strings treated as \code{NA} (see \href{https://specs.frictionlessdata.io/table-schema/#missing-values}{table-schema/missing-values}).}

\item{primaryKey}{(character) Name of primary key field(s) (see \href{https://specs.frictionlessdata.io/table-schema/#primary-key}{table-schema/primary-key}).}

\item{foreignKeys}{(list of named lists) See \code{\link{foreignKey}}.}

\item{...}{(name = value) Additional properties.}
}
\description{
A schema declaration for a tabular data resource.
}
\details{
TODO: Support writing schemas to file with a path argument.
}
\examples{
schema(
  fields = list(
    field(name = "id", type = "integer"),
    field(name = "value", type = "number")
  ),
  missingValues = c("", "NA"),
  primaryKey = "id",
  tableSchema = TRUE
)
}
\references{
\url{https://specs.frictionlessdata.io/table-schema/}
}
\seealso{
Other meta objects: \code{\link{constraints}},
  \code{\link{contributor}}, \code{\link{field}},
  \code{\link{foreignKey}}, \code{\link{package}},
  \code{\link{resource}}, \code{\link{source}}
}
