% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_by_name}
\alias{set_by_name}
\title{Set object metadata by name}
\usage{
set_by_name(x, fun, ...)
}
\arguments{
\item{x}{Object.}

\item{fun}{(function) Set function.}

\item{...}{(name = value) Element names of \code{x} equal to list of properties to set.}
}
\description{
Set object metadata by name
}
