% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\name{set_field}
\alias{set_field}
\alias{set_field<-}
\title{Set field metadata}
\usage{
set_field(.data, name, type, format, title, description, rdfType, constraints,
  unit, ...)

set_field(.data) <- value
}
\arguments{
\item{.data}{Object to modify (typically an atomic vector).}

\item{name}{(character) Field name.}

\item{type}{(character) Field type (see \href{https://specs.frictionlessdata.io/table-schema/#types-and-formats}{table-schema/types-and-formats}).}

\item{format}{(character) Field format (see \href{https://specs.frictionlessdata.io/table-schema/#types-and-formats}{table-schema/types-and-formats}).}

\item{title}{(character) Human-readable title.}

\item{description}{(character) Human-readable description.}

\item{rdfType}{(character) URI of an RDF class (see \href{https://specs.frictionlessdata.io/table-schema/#rich-types}{table-schema/rich-types}).}

\item{constraints}{(named list) See \code{\link{constraints}}.}

\item{unit}{(character) Unit of measurement.}

\item{...}{(name = value) Additional properties.}

\item{value}{(named list) Field metadata (typically a call to \code{\link{field}}).}
}
\description{
New properties are added and existing properties are replaced (or deleted if \code{NULL}). If set, properties override the defaults returned by \code{\link{get_field}}.
}
\examples{
x <- set_field(
 Sys.Date(),
 name = "date_created",
 type = "date",
 format = "\%Y-\%m-\%d"
)
get_field(x)
x <- Sys.Date()
set_field(x) <- field(name = "date_created", type = "date", format = "\%Y-\%m-\%d")
get_field(x)
}
\seealso{
Other field functions: \code{\link{field}},
  \code{\link{get_field}}, \code{\link{set_fields}}
}
