% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\name{set_fields}
\alias{set_fields}
\title{Set field metadata by name}
\usage{
set_fields(l, ...)
}
\arguments{
\item{l}{(list) List of objects to modify.}

\item{...}{(name = value) Field metadata (named list or call to \link{field}) to assign to the object of the given name.}
}
\description{
Set field metadata by name
}
\examples{
df <- set_fields(
  data.frame(id = 1L, value = 1.1),
  value = field(title = "Value"),
  id = field(title = "Unique identifier", constraints = constraints(unique = TRUE))
)
get_fields(df)
}
\seealso{
Other field functions: \code{\link{field}},
  \code{\link{get_field}}, \code{\link{set_field}}
}
