% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\name{set_package}
\alias{set_package}
\alias{set_package<-}
\title{Set package metadata}
\usage{
set_package(.data, name, title, description, homepage, id, profile, version,
  created, sources, contributors, licenses, keywords, image, resources, ...)

set_package(.data) <- value
}
\arguments{
\item{.data}{Object to modify (typically a list of data.frames).}

\item{name}{(character) Package name.}

\item{title}{(character) Human-readable title.}

\item{description}{(character) Human-readable description.}

\item{homepage}{(character) URL of the package homepage.}

\item{id}{(character) Globally unique identifier.}

\item{profile}{(character) Package profile (see \url{https://specs.frictionlessdata.io/profiles/}).}

\item{version}{(character) Version specification (see \url{http://semver.org/}).}

\item{created}{(character) ISO 8601 date and time of package creation (see \url{https://tools.ietf.org/html/rfc3339#section-5.6}).}

\item{sources}{(list) See \code{\link{source}}.}

\item{contributors}{(list) See \code{\link{contributor}}.}

\item{licenses}{(list) See \code{\link{license}}.}

\item{keywords}{(character) Keywords describing the package.}

\item{image}{(character) URL or path relative to the package directory (see \href{https://specs.frictionlessdata.io/data-resource/#url-or-path}{data-resource/url-or-path}).}

\item{resources}{(list) See \code{\link{resource}}.}

\item{...}{(name = value) Additional properties.}

\item{value}{(named list) Package metadata (typically a call to \code{\link{package}}).}
}
\description{
New properties are added and existing properties are replaced (or deleted if \code{NULL}). If set, properties override the defaults returned by \code{\link{get_package}}.
}
\examples{
x <- set_package(
  list(data = data.frame(id = 1L, value = 1.1)),
  title = "Example data package"
)
get_package(x)
x <- list(data = data.frame(id = 1L, value = 1.1))
set_package(x) <- package(title = "Example data package")
get_package(x, inline_data = FALSE)
}
\seealso{
Other package functions: \code{\link{package}}
}
