% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\name{set_resource}
\alias{set_resource}
\alias{set_resource<-}
\title{Set resource metadata}
\usage{
set_resource(.data, name, path, profile, title, description, format, mediatype,
  encoding, schema, bytes, hash, sources, licenses, ...)

set_resource(.data) <- value
}
\arguments{
\item{.data}{Object to modify (typically a data.frame).}

\item{name}{(character) Resource name.}

\item{path}{(character) URLs or paths relative to the data package directory (see \href{https://specs.frictionlessdata.io/data-resource/#path-data-in-files}{data-resource/path-data-in-files}).}

\item{profile}{(character) Custom resource profile (see \url{https://specs.frictionlessdata.io/profiles/}).}

\item{title}{(character) Human-readable title.}

\item{description}{(character) Human-readable description.}

\item{format}{(character) Output format (e.g. "csv", "json").}

\item{mediatype}{(character) Mediatype/mimetype of the output (e.g. "text/csv", "application/json").}

\item{encoding}{(character) Character encoding of the output (see \url{http://www.iana.org/assignments/character-sets/character-sets.xhtml}).}

\item{schema}{(named list) See \code{\link{schema}}.}

\item{bytes}{(integer) Size of the output file in bytes.}

\item{hash}{(character) MD5 hash of the output file.}

\item{sources}{(list) See \code{\link{source}}.}

\item{licenses}{(list) See \code{\link{license}}.}

\item{...}{(name = value) Additional properties.}

\item{value}{(named list) Resource metadata (typically a call to \code{\link{resource}}).}
}
\description{
New properties are added and existing properties are replaced (or deleted if \code{NULL}). If set, properties override the defaults returned by \code{\link{get_resource}}.
}
\examples{
x <- set_resource(
  data.frame(id = 1L, value = 1.1),
  title = "Example data",
  path = "data/example.csv"
)
get_resource(x)
x <- data.frame(id = 1L, value = 1.1)
set_resource(x) <- resource(title = "Example data", path = "data/example.csv")
get_resource(x)
}
\seealso{
Other resource functions: \code{\link{get_package}},
  \code{\link{get_resource}}, \code{\link{resource}},
  \code{\link{set_resources}}
}
