% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\name{set_resources}
\alias{set_resources}
\title{Set resource metadata by name}
\usage{
set_resources(l, ...)
}
\arguments{
\item{l}{(list) List of objects to modify.}

\item{...}{(name = value) Resource metadata (named list or call to \link{resource}) to assign to the object of the given name.}
}
\description{
Set resource metadata by name
}
\examples{
l <- set_resources(
  list(a = data.frame(), b = data.frame()),
  a = resource(path = "data/a.csv"),
  b = resource(path = "data/b.csv")
)
get_resources(l)
}
\seealso{
Other resource functions: \code{\link{get_package}},
  \code{\link{get_resource}}, \code{\link{resource}},
  \code{\link{set_resource}}
}
