% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpkg_release_gh.R
\name{stow_gh_release}
\alias{stow_gh_release}
\title{stow a github release asset created with dpkg_github_release()}
\usage{
stow_gh_release(owner, repo, dpkg, overwrite = FALSE)
}
\description{
stow a github release asset created with dpkg_github_release()
}
\examples{
stow_gh_release("cole-brokamp", "dpkg", "mtcars-v0.0.0.9000")
stow_gh_release("cole-brokamp", "dpkg", "mtcars-v0.0.0.9000", force = TRUE)
}
