% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpkg_release_gh.R
\name{stow_gh_release_asset}
\alias{stow_gh_release_asset}
\title{stow a github release asset created by dpkg_github_release()}
\usage{
stow_gh_release_asset(owner, repo, dpkg, force = FALSE)
}
\description{
stow a github release asset created by dpkg_github_release()
}
\examples{
stow_gh_release_asset("cole-brokamp", "dpkg", "mtcars-v0.0.0.9000")
stow_gh_release_asset("cole-brokamp", "dpkg", "mtcars-v0.0.0.9000", force = TRUE)
}
