% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stow.R
\name{stow_url}
\alias{stow_url}
\title{download a file to the \code{stow} R user directory}
\usage{
stow_url(url, overwrite = FALSE)
}
\arguments{
\item{url}{a URL string starting with \verb{http://} or \verb{https://}}

\item{overwrite}{logical; overwrite stowed file with the same name?}
}
\description{
download a file to the \code{stow} R user directory
}
\examples{
Sys.setenv(R_USER_DATA_DIR = tempfile("stow"))
stow_url("https://github.com/geomarker-io/appc/releases/download/v0.1.0/nei_2020.rds")
}
