% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\name{unset}
\alias{unset}
\title{Unset object metadata}
\usage{
unset(x, recursive = TRUE)
}
\arguments{
\item{x}{Object.}

\item{recursive}{(logical) Whether to process object elements.}
}
\description{
Removes \code{dpkg} and \code{units} attributes from an object.
}
\examples{
x = set_field(Sys.Date(), name = "test")
str(unset(x))
y = set_field(Sys.time(), name = "test")
str(unset(y))
z = set_field(TRUE, name = "test")
str(unset(z))
df <- set_resource(data.frame(x, y, z), name = "test")
str(unset(df))
l <- list(x, list(y, z))
str(unset(l))
\dontrun{
x = set_field(units2::as_units(1, "m"), name = "test")
str(unset(x))
}
}
