% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{write_csv}
\alias{write_csv}
\title{Write CSV}
\usage{
write_csv(x, file = NULL, na = "")
}
\arguments{
\item{x}{Object to write.}

\item{file}{(character) Path to file on disk. If \code{NULL}, a CSV string is returned.}

\item{na}{(character) String to use for missing values.}
}
\description{
Write CSV
}
\seealso{
Other formats: \code{\link{read_csv}},
  \code{\link{read_json}}, \code{\link{write_json}}
}
