% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_data}
\alias{write_data}
\title{Write package data}
\usage{
write_data(x, path = getwd())
}
\arguments{
\item{x}{Object.}

\item{path}{(character) Package directory.}
}
\description{
Write package data
}
\seealso{
Other package writers: \code{\link{write_meta}},
  \code{\link{write_package}}, \code{\link{write_resource}}
}
