% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_dpkg.R
\name{write_dpkg}
\alias{write_dpkg}
\title{write dpkg to disk}
\usage{
write_dpkg(x, dir)
}
\arguments{
\item{x}{a data package (\code{dpkg}) object}

\item{dir}{path to directory where dpkg parquet file will be written}
}
\value{
path to the written file, invisibly
}
\description{
write dpkg to disk
}
