% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{write_json}
\alias{write_json}
\title{Write JSON}
\usage{
write_json(x, file = NULL)
}
\arguments{
\item{x}{Object to write.}

\item{file}{(character) Path to file on disk. If \code{NULL}, a JSON string is returned.}
}
\description{
Write JSON
}
\seealso{
Other formats: \code{\link{read_csv}},
  \code{\link{read_json}}, \code{\link{write_csv}}
}
