% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_meta}
\alias{write_meta}
\title{Write package metadata}
\usage{
write_meta(x, name = NULL, path = getwd(), inline_data = TRUE)
}
\arguments{
\item{x}{Object.}

\item{name}{(character) Package name if not set explicitly.}

\item{path}{(character) Package directory.}

\item{inline_data}{(logical) Whether to include the contents of \code{x} as inline data (\code{$data}).}
}
\description{
Write package metadata
}
\seealso{
Other package writers: \code{\link{write_data}},
  \code{\link{write_package}}, \code{\link{write_resource}}
}
