% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_package}
\alias{write_package}
\title{Write package}
\usage{
write_package(x, name = NULL, path = getwd(), inline_data = TRUE)
}
\arguments{
\item{x}{Object.}

\item{name}{(character) Package name if not set explicitly.}

\item{path}{(character) Package directory.}

\item{inline_data}{(logical) Whether to include the contents of \code{x} as inline data (\code{$data}).}
}
\description{
Writes package data and metadata to disk using the following rules for each resource:
\itemize{
  \item format: If missing, \code{path} file extensions and \code{mediatype} are checked when writing. Only csv and json are supported.
  \item path: Only single local paths are supported, and if not set, data is saved in the metadata as an inline JSON object or a formatted string.
  \item encoding: Only the default "UTF-8" is supported.
}
}
\seealso{
Other package writers: \code{\link{as_markdown}},
  \code{\link{write_data}}, \code{\link{write_meta}},
  \code{\link{write_resource}}
}
