% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_resource}
\alias{write_resource}
\title{Write resource data}
\usage{
write_resource(x, path = NULL)
}
\arguments{
\item{x}{Object.}

\item{path}{(character) Package directory.}
}
\description{
Write resource data
}
\seealso{
Other package writers: \code{\link{as_markdown}},
  \code{\link{write_data}}, \code{\link{write_meta}},
  \code{\link{write_package}}
}
