% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpkg.R
\name{dpkg_meta}
\alias{dpkg_meta}
\title{get the metadata associated with a data package}
\usage{
dpkg_meta(x)
}
\arguments{
\item{x}{a dpkg object}
}
\value{
a list of metadata key value pairs
}
\description{
get the metadata associated with a data package
}
