% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbl_tbl.R
\name{get_labels}
\alias{get_labels}
\title{get labels}
\usage{
get_labels(x, missing_label = "deparse")
}
\arguments{
\item{x}{a \code{lbl_tbl} object (see \code{as_lbl_tbl()})}

\item{missing_label}{what to return in the case of no label attribute: \code{"na"} returns missing values,
\code{"deparse"} substitutes and deparses the code used to define \code{x} into a character string, and
\code{"as_label"} tries to convert the code supplied as \code{x} to a label with \code{rlang::as_label()}}
}
\value{
character vector of labels; \code{NA} values are used if a label does not exist
}
\description{
Get the labels of labeled vectors in a labeled table
}
\examples{
d <-
  tibble::tibble(
    id = labels(letters[1:5]),
    letters = as_lbl_vec(letters[1:5], "The Alphabet"),
    numbers = as_lbl_vec(1:5, "the numbers"),
    logicals = as_lbl_vec(c(TRUE, TRUE, FALSE, TRUE, FALSE), "is foofy")
  )
d <- as_lbl_tbl(d)
get_labels(d)
}
