% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_dpkg.R
\name{read_dpkg}
\alias{read_dpkg}
\title{read dpkg from disk}
\usage{
read_dpkg(x)
}
\arguments{
\item{x}{path to data package (\code{.parquet} file) on disk}
}
\value{
a dpkg object
}
\description{
read dpkg from disk
}
