% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stow.R
\name{stow}
\alias{stow}
\title{stow}
\usage{
stow(uri, overwrite = FALSE)
}
\arguments{
\item{uri}{character string universal resource identifier; currently, must begin
with \verb{http://}, \verb{https://}, or \verb{gh://}}

\item{overwrite}{logical; re-download the remote file even though
a local file with the same name exists?}
}
\value{
path to the stowed file
}
\description{
Stow a remote file locally, caching it across all of a user's
R sessions and projects.
}
\details{
Use stow to abstract away the process of downloading a file
or a GitHub release asset to a user's data directory, only
downloading files that have not already been downloaded.

Supported URI prefixes include:
\itemize{
\item \verb{https://}, \verb{http://}: download from a file
\item \verb{gh://}: download a github release asset, formatted as \verb{gh://owner/repo/name}
}

Stow downloads files to the users data directory; see \code{?tools::R_user_dir}.
Specify an alternative download location by setting the \code{R_USER_DATA_DIR}
environment variable.
The stow cache works by name only; that is, if a file with the same URI
has already been downloaded once, it will not be re-downloaded again
(unless \code{overwrite = TRUE}).
}
\examples{
# get by using URL
stow("https://github.com/geomarker-io/appc/releases/download/v0.1.0/nei_2020.rds")

# will be faster (even in later R sessions) next time
stow("https://github.com/geomarker-io/appc/releases/download/v0.1.0/nei_2020.rds") |>
  readRDS()

# get a data package parquet file created with dpkg_gh_release()
stow("gh://cole-brokamp/dpkg/cagis_parcels-v0.1.0") |>
  read_dpkg()
}
