% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stow.R
\name{stow}
\alias{stow}
\title{download a file to the \code{stow} R user directory}
\usage{
stow(url, overwrite = FALSE)
}
\arguments{
\item{url}{a URL string starting with \verb{http://} or \verb{https://}}

\item{overwrite}{logical; overwrite stowed file with the same name?}
}
\description{
Downloaded files are available across all of an R user’s sessions and projects.
If a file with the same name has already been downloaded once,
it will not be re-downloaded again, unless \code{force = TRUE}.
Specify an alternative download location by setting the \code{R_USER_DATA_DIR}
environment variable; see \code{?tools::R_user_dir}.
}
\examples{
Sys.setenv(R_USER_DATA_DIR = tempfile("stow"))
stow("https://github.com/geomarker-io/appc/releases/download/v0.1.0/nei_2020.rds")
}
